+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [6.4.0] - 2019-10-04
+
+### Added
+- Listing SVF image assets
+- Parsing and querying SVF property database
+
 ## [6.3.0] - 2019-10-03
 
 ### Added